moved some blas1_trees operations to the generic trees header
