+++ include/operations/blas1_trees.hpp
@@ -30,240 +30,10 @@
 #include <vector>
 
 #include <operations/blas_operators.hpp>
+#include <operations/blas_trees.hpp>
 #include <views/view_sycl.hpp>
 
 namespace blas {
-namespace internal {
-
-/*! DetectScalar.
- * @brief Class specialization used to detect scalar values in ScalarOp nodes.
- * When the value is not an integral basic type,
- * it is assumed to be a vector and the first value
- * is used.
- */
-template <typename T>
-struct DetectScalar {
-  static typename T::value_type get_scalar(T &opSCL) { return opSCL.eval(0); }
-};
-
-/*! DetectScalar.
- * @brief See Detect Scalar.
- */
-template <>
-struct DetectScalar<int> {
-  using T = int;
-  static T get_scalar(T &scalar) { return scalar; }
-};
-
-/*! DetectScalar.
- * @brief See Detect Scalar.
- */
-template <>
-struct DetectScalar<float> {
-  using T = float;
-  static T get_scalar(T &scalar) { return scalar; }
-};
-
-/*! DetectScalar.
- * @brief See Detect Scalar.
-*/
-template <>
-struct DetectScalar<double> {
-  using T = double;
-  static T get_scalar(T &scalar) { return scalar; }
-};
-
-/*! DetectScalar.
- * @brief See Detect Scalar.
-*/
-template <>
-struct DetectScalar<std::complex<float>> {
-  using T = std::complex<float>;
-  static T get_scalar(T &scalar) { return scalar; }
-};
-
-/*! DetectScalar.
- * @brief See Detect Scalar.
-*/
-template <>
-struct DetectScalar<std::complex<double>> {
-  using T = std::complex<double>;
-  static T get_scalar(T &scalar) { return scalar; }
-};
-
-/*! get_scalar.
- * @brief Template autodecuction function for DetectScalar.
-*/
-template <typename T>
-auto get_scalar(T &scl) -> decltype(DetectScalar<T>::get_scalar(scl)) {
-  return DetectScalar<T>::get_scalar(scl);
-}
-}  // namespace internal
-
-/** Join.
- * @brief Joins both sides of the expression in the single kernel.
- */
-template <class LHS, class RHS>
-struct Join {
-  LHS l;
-  RHS r;
-
-  using value_type = typename RHS::value_type;
-
-  Join(LHS &_l, RHS _r) : l(_l), r(_r){};
-
-  // PROBLEM: Only the RHS size is considered. If LHS size is different??
-  // If it is smaller, eval function will crash
-  size_t getSize() { return r.getSize(); }
-
-  value_type eval(size_t i) {
-    l.eval(i);
-    return r.eval(i);
-  }
-
-  value_type eval(cl::sycl::nd_item<1> ndItem) {
-    return eval(ndItem.get_global(0));
-  }
-};
-
-/** Assign.
- */
-template <class LHS, class RHS>
-struct Assign {
-  LHS l;
-  RHS r;
-
-  using value_type = typename RHS::value_type;
-
-  Assign(LHS &_l, RHS _r) : l(_l), r(_r){};
-
-  // PROBLEM: Only the RHS size is considered. If LHS size is different??
-  // If it is smaller, eval function will crash
-  size_t getSize() { return r.getSize(); }
-
-  value_type eval(size_t i) {
-    auto val = l.eval(i) = r.eval(i);
-    return val;
-  }
-
-  value_type eval(cl::sycl::nd_item<1> ndItem) {
-    return eval(ndItem.get_global(0));
-  }
-};
-
-/*! DoubleAssign.
- */
-template <class LHS1, class LHS2, class RHS1, class RHS2>
-struct DobleAssign {
-  LHS1 l1;
-  LHS2 l2;
-  RHS1 r1;
-  RHS2 r2;
-
- public:
-  using value_type = typename RHS1::value_type;
-
-  DobleAssign(LHS1 &_l1, LHS2 &_l2, RHS1 _r1, RHS2 _r2)
-      : l1(_l1), l2(_l2), r1(_r1), r2(_r2){};
-
-  // PROBLEM: Only the RHS size is considered. If LHS size is different??
-  // If it is smaller, eval function will crash
-  size_t getSize() { return r2.getSize(); }
-
-  value_type eval(size_t i) {
-    auto val1 = r1.eval(i);
-    auto val2 = r2.eval(i);
-    l1.eval(i) = val1;
-    l2.eval(i) = val2;
-    return val1;
-  }
-
-  value_type eval(cl::sycl::nd_item<1> ndItem) {
-    return eval(ndItem.get_global(0));
-  }
-};
-
-/*!ScalarOp.
- * @brief Implements an scalar operation.
- * (e.g alpha OP x, with alpha scalar and x vector)
- */
-template <typename Operator, typename SCL, typename RHS>
-struct ScalarOp {
-  using value_type = typename RHS::value_type;
-  SCL scl;
-  RHS r;
-
-  ScalarOp(SCL _scl, RHS &_r) : scl(_scl), r(_r){};
-
-  size_t getSize() { return r.getSize(); }
-  value_type eval(size_t i) {
-    return Operator::eval(internal::get_scalar(scl), r.eval(i));
-  }
-
-  value_type eval(cl::sycl::nd_item<1> ndItem) {
-    return eval(ndItem.get_global(0));
-  }
-};
-
-/*! UnaryOp.
- * Implements a Unary Operation ( Op(z), e.g. z++), with z a vector.
- */
-template <typename Operator, typename RHS>
-struct UnaryOp {
-  using value_type = typename RHS::value_type;
-  RHS r;
-
-  UnaryOp(RHS &_r) : r(_r){};
-
-  size_t getSize() { return r.getSize(); }
-
-  value_type eval(size_t i) { return Operator::eval(r.eval(i)); }
-
-  value_type eval(cl::sycl::nd_item<1> ndItem) {
-    return eval(ndItem.get_global(0));
-  }
-};
-
-/*! BinaryOp.
- * @brief Implements a Binary Operation (x OP z) with x and z vectors.
- */
-template <typename Operator, typename LHS, typename RHS>
-struct BinaryOp {
-  using value_type = typename RHS::value_type;
-  LHS l;
-  RHS r;
-
-  BinaryOp(LHS &_l, RHS &_r) : l(_l), r(_r){};
-
-  // PROBLEM: Only the RHS size is considered. If LHS size is different??
-  // If it is smaller, eval function will crash
-  size_t getSize() { return r.getSize(); }
-
-  value_type eval(size_t i) { return Operator::eval(l.eval(i), r.eval(i)); }
-
-  value_type eval(cl::sycl::nd_item<1> ndItem) {
-    return eval(ndItem.get_global(0));
-  }
-};
-
-/*! TupleOp.
- * @brief Implements a Tuple Operation (map (\x -> [i, x]) vector).
- */
-template <typename RHS>
-struct TupleOp {
-  using value_type = IndVal<typename RHS::value_type>;
-  RHS r;
-
-  TupleOp(RHS &_r) : r(_r) {}
-
-  size_t getSize() { return r.getSize(); }
-
-  value_type eval(size_t i) { return value_type(i, r.eval(i)); }
-
-  value_type eval(cl::sycl::nd_item<1> ndItem) {
-    return eval(ndItem.get_global(0));
-  }
-};
 
 /*! AssignReduction.
  * @brief Implements the reduction operation for assignments (in the form y = x)
